chase.c, cl_input.c, cl_parse.c, client.h, common.c, common.h, console.h, cvar.h, draw.h, gl_draw.c, gl_fog.c, gl_mesh.c, gl_model.c, gl_model.h, gl_rmain.c, gl_rmisc.c, gl_screen.c, gl_sky.c, gl_texmgr.c, glquake.h, host.c, keys.c, keys.h, main.c, menu.c, menu.h, pr_cmds.c, quakedef.h, r_alias.c, r_brush.c, r_part.c, r_sprite.c, r_world.c, sbar.c, sbar.h, screen.h, snd_dma.c, snd_mem.c, snd_mix.c, sv_main.c, sys_sdl.c, vid.h, view.h, world.c, world.h:  Loads of warning fixes about missing function prototypes, missing parens around &, missing braces leading to ambiguous else statements and unused and uninitialized variables. There are still a couple of unitialised variables here and there, but not much. The warnings about strict aliasing violations need taking care of.
git-svn-id: svn://svn.code.sf.net/p/quakespasm/code/trunk/quakespasm@21 af15c1b1-3010-417e-b628-4374ebc0bcbd

+++ Quake/common.c
@@ -1143,7 +1143,7 @@ void COM_InitArgv (int argc, char **argv)
 			break;
 	}
 
-	if (com_cmdline[n-1] = ' ')
+	if (n > 0 && com_cmdline[n-1] == ' ')
 		com_cmdline[n-1] = 0; //johnfitz -- kill the trailing space
 
 	Con_Printf("\nCommand line: %s",com_cmdline);
